requirements.txt
appdirs==1.4.3
certifi==2018.1.18
chardet==3.0.4
colorclass==2.2.0
dj-database-url==0.4.2
Django==2.0.2
django-appconf==1.0.2
django-compressor==2.2
django-crispy-forms==1.7.0
django-filter==1.1.0
django-filters==0.2.1
django-jet==1.0.7
djangorestframework==3.7.7
docopt==0.6.2
gunicorn==19.7.1
idna==2.6
olefile==0.45.1
packaging==16.8
Pillow==5.0.0
pip-upgrader==1.4.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
style==1.1.3
terminaltables==3.1.0
update==0.0.1
urllib3==1.22
